chore: Move area related files to their own package in :data module
