+++ data/src/main/kotlin/com/routesearch/data/DataModule.kt
@@ -1,5 +1,7 @@
 package com.routesearch.data
 
+import com.routesearch.data.area.AreaRepository
+import com.routesearch.data.area.DefaultAreaRepository
 import com.routesearch.data.local.localDataModule
 import com.routesearch.data.remote.remoteDataModule
 import org.koin.core.module.dsl.singleOf
