+++ data/src/main/kotlin/com/routesearch/data/area/AreaRepository.kt
@@ -1,4 +1,4 @@
-package com.routesearch.data
+package com.routesearch.data.area
 
 interface AreaRepository {
